Start implementation of pushdown transducer
Implements the bare boned of the pushdown transducer but doesn't yet
implement the merge function. This commit also slightly changes the API
of the state map so that all required operations can be completed in a
single API call so that iterator invalidation is not a problem. Finally
this commit also includes the first version of the transition representation.